Move profile card to separated component
